Add login method to ApiService
Added a new `login` method to the `ApiService` class to handle user login requests. This includes updating the `api_service.dart` file to import the `LoginRequest` model and defining the `login` method. Additionally, the generated file `api_service.g.dart` was updated to include the implementation of the `login` method.