Update BestSeller and BestSellingGridView to use dynamic data
Updated the BestSeller widget to accept a `CategoriesDetailsEntity` object and use its properties for the image, name, and price. Modified the BestSellingGridView to accept a list of `CategoriesDetailsEntity` objects and pass each item to the BestSeller widget. Also, updated the BestSellingHeader to pass the data to the all_productis route.